+++ lib/reconcile/rabbitmq.go
@@ -408,6 +408,21 @@ func getRabbitMQPodSpec(statefulSetName, dataVolumeName string, cr *apiv1alpha2.
 		},
 	}
 
+	// do we want priorities?
+	if cr.Spec.Features.AstartePodPriorities.IsEnabled() {
+		//is a priorityClass specified in the Astarte CR?
+		switch cr.Spec.RabbitMQ.AstarteGenericClusteredResource.PriorityClass {
+		case highPriority:
+			ps.PriorityClassName = AstarteHighPriorityName
+		case midPriority:
+			ps.PriorityClassName = AstarteMidPriorityName
+		case lowPriority:
+			ps.PriorityClassName = AstarteLowPriorityName
+		default:
+			ps.PriorityClassName = AstarteHighPriorityName
+		}
+	}
+
 	return ps
 }