+++ lib/reconcile/vernemq.go
@@ -342,6 +342,21 @@ func getVerneMQPodSpec(statefulSetName, dataVolumeName string, cr *apiv1alpha2.A
 		Volumes: getVerneMQVolumes(cr),
 	}
 
+	// do we want priorities?
+	if cr.Spec.Features.AstartePodPriorities.IsEnabled() {
+		//is a priorityClass specified in the Astarte CR?
+		switch cr.Spec.VerneMQ.AstarteGenericClusteredResource.PriorityClass {
+		case highPriority:
+			ps.PriorityClassName = AstarteHighPriorityName
+		case midPriority:
+			ps.PriorityClassName = AstarteMidPriorityName
+		case lowPriority:
+			ps.PriorityClassName = AstarteLowPriorityName
+		default:
+			ps.PriorityClassName = AstarteHighPriorityName
+		}
+	}
+
 	return ps
 }